Remove image from small screens
It's just distracting on a mobile device. Plus it strikes the text in slightly awkward positions and ends too early for the text I have right now. So I'm seeing what happens if I get rid of it. (I may switch to having it in a fixed position instead.) Now let's see if I got the query right.